Added missing tests for Phnx.Tests and reduced fragility of ToOrdinal, FromOrdinal and Remove
