+++ src/Phnx/Phnx.Tests/Extensions/StringExtensionsTests/PascalCamelCaseTests/From.cs
@@ -6,111 +6,110 @@ namespace Phnx.Tests.Extensions.StringExtensionsTests.PascalCamelCaseTests
     public class From
     {
         [Test]
-        public void ToCamelCase_WithNullString_ThrowsArgumentNullException()
+        public void FromCamelCase_WithNullString_ThrowsArgumentNullException()
         {
             string testString = null;
 
-            Assert.Throws<ArgumentNullException>(() => testString.ToCamelCase(false));
-            Assert.Throws<ArgumentNullException>(() => testString.ToPascalCase());
+            Assert.Throws<ArgumentNullException>(() => testString.FromCamelAndPascalCase(false));
         }
 
         [Test]
-        public void ToCamelCase_WithEmptyString_ReturnsEmptyString()
+        public void FromCamelCase_WithEmptyString_ReturnsEmptyString()
         {
             string testString = string.Empty;
 
-            string result = testString.ToCamelCase(false);
+            string result = testString.FromCamelAndPascalCase(false);
 
             Assert.AreEqual(testString, result);
 
-            result = testString.ToPascalCase();
+            result = testString.FromCamelAndPascalCase(true);
 
             Assert.AreEqual(testString, result);
         }
 
         [Test]
-        public void ToCamelCase_WithSingleWord_ReturnsSingleWord()
+        public void FromCamelCase_WithSingleWord_ReturnsSingleWord()
         {
             string testString = "asdf";
 
-            string result = testString.ToCamelCase(false);
+            string result = testString.FromCamelAndPascalCase(false);
 
             Assert.AreEqual(testString, result);
         }
 
         [Test]
-        public void ToPascalCase_WithSingleWord_ReturnsSingleWordWithFirstCharUpperCase()
+        public void FromCamelCase_WithSingleWordAndCapitalWords_ReturnsSingleWordStartingWithCapital()
         {
             string testString = "asdf";
             string expected = "Asdf";
 
-            string result = testString.ToPascalCase();
+            string result = testString.FromCamelAndPascalCase(true);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void ToCamelCase_WithSentence_ReturnsCamelCase()
+        public void FromCamelCase_WithSentenceAndLowercaseWords_ReturnsSentence()
         {
-            string testString = "Pack my box with five dozen liquor jugs";
-            string expected = "packMyBoxWithFiveDozenLiquorJugs";
+            string testString = "packMyBoxWithFiveDozenLiquorJugs";
+            string expected = "pack my box with five dozen liquor jugs";
 
-            string result = testString.ToCamelCase(false);
+            string result = testString.FromCamelAndPascalCase(false);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void ToPascalCase_WithSentence_ReturnsCamelCase()
+        public void FromPascalCase_WithSentenceAndUppercaseWords_ReturnsSentence()
         {
-            string testString = "Pack my box with five dozen liquor jugs";
-            string expected = "PackMyBoxWithFiveDozenLiquorJugs";
+            string testString = "PackMyBoxWithFiveDozenLiquorJugs";
+            string expected = "Pack My Box With Five Dozen Liquor Jugs";
 
-            string result = testString.ToPascalCase();
+            string result = testString.FromCamelAndPascalCase(true);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void ToCamelCase_WithSentenceContainingDoubleSpace_IgnoresDoubleSpaceAndReturnsCamelCase()
+        public void FromCamelCase_WithNumbersInSentenceAndLowercaseWords_ReturnsSentence()
         {
-            string testString = "Pack  my  box  with  five  dozen  liquor  jugs";
-            string expected = "packMyBoxWithFiveDozenLiquorJugs";
+            string testString = "packMyBoxWith54DozenLiquorJugs";
+            string expected = "pack my box with 54 dozen liquor jugs";
 
-            string result = testString.ToCamelCase(false);
+            string result = testString.FromCamelAndPascalCase(false);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void ToPascalCase_WithSentenceContainingDoubleSpace_IgnoresDoubleSpaceAndReturnsPascalCase()
+        public void FromPascalCase_WithNumbersInSentenceAndUppercaseWords_ReturnsSentence()
         {
-            string testString = "Pack  my  box  with  five  dozen  liquor  jugs";
-            string expected = "PackMyBoxWithFiveDozenLiquorJugs";
+            string testString = "PackMyBoxWith54DozenLiquorJugs";
+            string expected = "Pack My Box With 54 Dozen Liquor Jugs";
 
-            string result = testString.ToPascalCase();
+            string result = testString.FromCamelAndPascalCase(true);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void ToCamelCase_WithNumbersInSentence_ReturnsCamelCase()
+        public void FromCamelCase_WithAcronym_ReturnsSentence()
         {
-            string testString = "Pack my box with 54 dozen liquor jugs";
-            string expected = "packMyBoxWith54DozenLiquorJugs";
+            string testString = "thisIsAnIDForAUser";
+            string expected = "this is an ID for a user";
 
-            string result = testString.ToCamelCase(false);
+            string result = testString.FromCamelAndPascalCase(false);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void ToPascalCase_WithNumbersInSentence_ReturnsCamelCase()
+        public void FromPascalCase_WithAcronym_ReturnsSentence()
         {
-            string testString = "Pack my box with 54 dozen liquor jugs";
-            string expected = "PackMyBoxWith54DozenLiquorJugs";
+            string testString = "thisIsAnIDForAUser";
+            string expected = "This Is An ID For A User";
 
-            string result = testString.ToPascalCase();
+            string result = testString.FromCamelAndPascalCase(true);
 
             Assert.AreEqual(expected, result);
         }
