+++ src/Phnx/Phnx.Tests/Extensions/StringExtensionsTests/PascalCamelCaseTests/To.cs
@@ -6,110 +6,111 @@ namespace Phnx.Tests.Extensions.StringExtensionsTests.PascalCamelCaseTests
     public class To
     {
         [Test]
-        public void FromCamelCase_WithNullString_ThrowsArgumentNullException()
+        public void ToCamelCase_WithNullString_ThrowsArgumentNullException()
         {
             string testString = null;
 
-            Assert.Throws<ArgumentNullException>(() => testString.FromCamelAndPascalCase(false));
+            Assert.Throws<ArgumentNullException>(() => testString.ToCamelCase(false));
+            Assert.Throws<ArgumentNullException>(() => testString.ToPascalCase());
         }
 
         [Test]
-        public void FromCamelCase_WithEmptyString_ReturnsEmptyString()
+        public void ToCamelCase_WithEmptyString_ReturnsEmptyString()
         {
             string testString = string.Empty;
 
-            string result = testString.FromCamelAndPascalCase(false);
+            string result = testString.ToCamelCase(false);
 
             Assert.AreEqual(testString, result);
 
-            result = testString.FromCamelAndPascalCase(true);
+            result = testString.ToPascalCase();
 
             Assert.AreEqual(testString, result);
         }
 
         [Test]
-        public void FromCamelCase_WithSingleWord_ReturnsSingleWord()
+        public void ToCamelCase_WithSingleWord_ReturnsSingleWord()
         {
             string testString = "asdf";
 
-            string result = testString.FromCamelAndPascalCase(false);
+            string result = testString.ToCamelCase(false);
 
             Assert.AreEqual(testString, result);
         }
 
         [Test]
-        public void FromCamelCase_WithSingleWordAndCapitalWords_ReturnsSingleWordStartingWithCapital()
+        public void ToPascalCase_WithSingleWord_ReturnsSingleWordWithFirstCharUpperCase()
         {
             string testString = "asdf";
             string expected = "Asdf";
 
-            string result = testString.FromCamelAndPascalCase(true);
+            string result = testString.ToPascalCase();
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void FromCamelCase_WithSentenceAndLowercaseWords_ReturnsSentence()
+        public void ToCamelCase_WithSentence_ReturnsCamelCase()
         {
-            string testString = "packMyBoxWithFiveDozenLiquorJugs";
-            string expected = "pack my box with five dozen liquor jugs";
+            string testString = "Pack my box with five dozen liquor jugs";
+            string expected = "packMyBoxWithFiveDozenLiquorJugs";
 
-            string result = testString.FromCamelAndPascalCase(false);
+            string result = testString.ToCamelCase(false);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void FromPascalCase_WithSentenceAndUppercaseWords_ReturnsSentence()
+        public void ToPascalCase_WithSentence_ReturnsCamelCase()
         {
-            string testString = "PackMyBoxWithFiveDozenLiquorJugs";
-            string expected = "Pack My Box With Five Dozen Liquor Jugs";
+            string testString = "Pack my box with five dozen liquor jugs";
+            string expected = "PackMyBoxWithFiveDozenLiquorJugs";
 
-            string result = testString.FromCamelAndPascalCase(true);
+            string result = testString.ToPascalCase();
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void FromCamelCase_WithNumbersInSentenceAndLowercaseWords_ReturnsSentence()
+        public void ToCamelCase_WithSentenceContainingDoubleSpace_IgnoresDoubleSpaceAndReturnsCamelCase()
         {
-            string testString = "packMyBoxWith54DozenLiquorJugs";
-            string expected = "pack my box with 54 dozen liquor jugs";
+            string testString = "Pack  my  box  with  five  dozen  liquor  jugs";
+            string expected = "packMyBoxWithFiveDozenLiquorJugs";
 
-            string result = testString.FromCamelAndPascalCase(false);
+            string result = testString.ToCamelCase(false);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void FromPascalCase_WithNumbersInSentenceAndUppercaseWords_ReturnsSentence()
+        public void ToPascalCase_WithSentenceContainingDoubleSpace_IgnoresDoubleSpaceAndReturnsPascalCase()
         {
-            string testString = "PackMyBoxWith54DozenLiquorJugs";
-            string expected = "Pack My Box With 54 Dozen Liquor Jugs";
+            string testString = "Pack  my  box  with  five  dozen  liquor  jugs";
+            string expected = "PackMyBoxWithFiveDozenLiquorJugs";
 
-            string result = testString.FromCamelAndPascalCase(true);
+            string result = testString.ToPascalCase();
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void FromCamelCase_WithAcronym_ReturnsSentence()
+        public void ToCamelCase_WithNumbersInSentence_ReturnsCamelCase()
         {
-            string testString = "thisIsAnIDForAUser";
-            string expected = "this is an ID for a user";
+            string testString = "Pack my box with 54 dozen liquor jugs";
+            string expected = "packMyBoxWith54DozenLiquorJugs";
 
-            string result = testString.FromCamelAndPascalCase(false);
+            string result = testString.ToCamelCase(false);
 
             Assert.AreEqual(expected, result);
         }
 
         [Test]
-        public void FromPascalCase_WithAcronym_ReturnsSentence()
+        public void ToPascalCase_WithNumbersInSentence_ReturnsCamelCase()
         {
-            string testString = "thisIsAnIDForAUser";
-            string expected = "This Is An ID For A User";
+            string testString = "Pack my box with 54 dozen liquor jugs";
+            string expected = "PackMyBoxWith54DozenLiquorJugs";
 
-            string result = testString.FromCamelAndPascalCase(true);
+            string result = testString.ToPascalCase();
 
             Assert.AreEqual(expected, result);
         }
